project_web_naves/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
djangorestframework==3.15.2
numpy==1.24.4
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
